Ensure that the MPS builds with Clang 10.

(cherry picked from commit 8797baba7b10c99876d6bd3a19ad7f23db6b5217)

+++ code/pool.c
@@ -231,7 +231,7 @@ Res PoolAlloc(Addr *pReturn, Pool pool, Size size)
 
   /* All PoolAllocs should advance the allocation clock, so we count */
   /* it all in the fillMutatorSize field. */
-  ArenaGlobals(PoolArena(pool))->fillMutatorSize += size;
+  ArenaGlobals(PoolArena(pool))->fillMutatorSize += (double)size;
 
   EVENT3(PoolAlloc, pool, *pReturn, size);
